Update first time using web scraping with Python on wallstreet.cn.py
Since I am right now working in Ping An Technology. And I am using company's computer to try to scrape latest news concerning all of the subsidiary company of Ping An Group. However, there are limited python libraries available on my company's computer. I decided to simply use the library(requests) to finish this task. The website Wallstreet.cn is my first approach in the field of web scraping and will not be the last target.